+++ Zojae031/CodingTest/Debug/CodingTest.log
@@ -1,2 +1,2 @@
-﻿  2875.cpp
-  CodingTest.vcxproj -> C:\Users\USER\source\repos\CodingTest\Debug\CodingTest.exe
+﻿  14717.cpp
+LINK : fatal error LNK1168: C:\Users\USER\Desktop\git\Sejong Algorithm\Zojae031\Debug\CodingTest.exe을(를) 쓰기용으로 열 수 없습니다.
