+++ Zojae031/CodingTest/Debug/CodingTest.tlog/CodingTest.lastbuildstate
@@ -1,2 +1,2 @@
 #TargetFrameworkVersion=v4.0:PlatformToolSet=v141:EnableManagedIncrementalBuild=false:VCToolArchitecture=Native32Bit:WindowsTargetPlatformVersion=10.0.17763.0
-Debug|Win32|C:\Users\USER\source\repos\CodingTest\|
+Debug|Win32|C:\Users\USER\Desktop\git\Sejong Algorithm\Zojae031\|
